google map add

+++ db/migrate/20151029085812_create_items.rb
@@ -5,8 +5,8 @@ class CreateItems < ActiveRecord::Migration
       t.string :category 
       t.text :e_key , :null => false
       t.binary :pic_data, limit:10.megabyte
-      t.integer :pos_x , :precision => 9, :scale => 6
-      t.integer :pos_y , :precision => 9, :scale => 6
+      t.float :pos_x , :precision => 9, :scale => 6
+      t.float :pos_y , :precision => 9, :scale => 6
       t.string :f_name , :null => false
       t.text :f_id , :null => false
       t.datetime :f_date
